requirements.txt
streamlit==1.31.0
transformers==4.38.0
torch==2.2.0
vaderSentiment==3.3.2
newsapi-python==0.2.7
yfinance==0.2.36
plotly==5.18.0
pandas==2.2.0
numpy==1.26.0
requests==2.31.0
python-dotenv==1.0.0
pyngrok==7.0.0
